errors and simple validations

+++ server/src/schema/resolvers/Company.ts
@@ -1,4 +1,6 @@
-import { PrismaClient } from "@prisma/client";
+import { PrismaClient, Role } from "@prisma/client";
+import { AuthenticationError, UserInputError } from "apollo-server-express";
+import { Request } from "express";
 import {
 	CreateCompanyInput,
 	UpdateCompanyInput,
@@ -8,22 +10,55 @@ const prisma = new PrismaClient({ log: ["query"] });
 
 const companyResolvers = {
 	Query: {
-		companies: () => {
+		companies: (parent: any, args: unknown, { req }: { req: Request }) => {
+			if (!req.session.user?.role.includes(Role.ADMIN)) {
+				throw new AuthenticationError(
+					"You don't have permissions for this action"
+				);
+			}
 			return prisma.$queryRaw`SELECT * FROM public."Company";`;
 		},
 	},
 	Mutation: {
-		createCompany: (parent: any, { input }: { input: CreateCompanyInput }) => {
+		createCompany: (
+			parent: any,
+			{ input }: { input: CreateCompanyInput },
+			{ req }: { req: Request }
+		) => {
+			if (!req.session.user?.role.includes(Role.ADMIN)) {
+				throw new AuthenticationError(
+					"You don't have permissions for this action"
+				);
+			}
+			if (input.name.length < 3) throw new UserInputError("Name is too short");
 			return prisma.company.create({ data: input });
 		},
-		updateCompany: (parent: any, { input }: { input: UpdateCompanyInput }) => {
+		updateCompany: (
+			parent: any,
+			{ input }: { input: UpdateCompanyInput },
+			{ req }: { req: Request }
+		) => {
+			if (!req.session.user?.role.includes(Role.ADMIN)) {
+				throw new AuthenticationError(
+					"You don't have permissions for this action"
+				);
+			}
 			const { id, name } = input;
 			return prisma.company.update({
 				where: { id: id },
 				data: { name },
 			});
 		},
-		deleteCompany: async (parent: any, { id }: { id: number }) => {
+		deleteCompany: async (
+			parent: any,
+			{ id }: { id: number },
+			{ req }: { req: Request }
+		) => {
+			if (!req.session.user?.role.includes(Role.ADMIN)) {
+				throw new AuthenticationError(
+					"You don't have permissions for this action"
+				);
+			}
 			const data = await prisma.company.delete({ where: { id: id } });
 			if (data) return true;
 			return false;
