+++ server/src/schema/resolvers/Product.ts
@@ -1,10 +1,12 @@
-import { PrismaClient } from "@prisma/client";
+import { PrismaClient, Role } from "@prisma/client";
 import GraphQLJSON from "graphql-type-json";
 import {
 	CreateProductInput,
 	UpdateProductInput,
 } from "../../generated/graphql";
+import { Request } from "express";
 import { v2 as cloudinary } from "cloudinary";
+import { AuthenticationError, UserInputError } from "apollo-server-express";
 
 const prisma = new PrismaClient({ log: ["query"] });
 
@@ -12,16 +14,22 @@ const productResolvers = {
 	JSON: GraphQLJSON,
 	Query: {
 		products: () => {
-			return prisma.product.findMany({ include: { company: true } });
+			return prisma.product.findMany();
+			// return prisma.product.findMany({ include: { company: true } });
 			// return prisma.$queryRaw`SELECT * FROM public."Product";`;
 		},
 	},
 	Mutation: {
 		createProduct: async (
 			parent: any,
-			{ input }: { input: CreateProductInput }
+			{ input }: { input: CreateProductInput },
+			{ req }: { req: Request }
 		) => {
-			// 0.5. check for user role (same in company,category) - middleware
+			if (!req.session.user?.role.includes(Role.ADMIN)) {
+				throw new AuthenticationError(
+					"You don't have permissions for this action"
+				);
+			}
 			const {
 				category_id,
 				company_id,
@@ -32,6 +40,7 @@ const productResolvers = {
 				shipping_cost,
 				discount,
 			} = input;
+			if (name.length < 3) throw new UserInputError("Name is too short");
 
 			const arrOfImgs = input.img_id.map((item, i) => {
 				return { img_id: item, img_src: input.img_src[i] };
@@ -49,6 +58,12 @@ const productResolvers = {
 				images: arrOfImgs,
 			};
 
+			const product = await prisma.product
+				.create({ data: newProduct })
+				.catch(() => {
+					input.img_id.forEach((id) => cloudinary.uploader.destroy(id!));
+				});
+
 			//create connection between company and category
 			await prisma.category.update({
 				where: { id: category_id },
@@ -56,12 +71,18 @@ const productResolvers = {
 					companies: { connect: { id: company_id } },
 				},
 			});
-
-			return prisma.product.create({ data: newProduct }).catch(() => {
-				input.img_id.forEach((id) => cloudinary.uploader.destroy(id!));
-			});
+			return product;
 		},
-		updateProduct: (parent: any, { input }: { input: UpdateProductInput }) => {
+		updateProduct: async (
+			parent: any,
+			{ input }: { input: UpdateProductInput },
+			{ req }: { req: Request }
+		) => {
+			if (!req.session.user?.role.includes(Role.ADMIN)) {
+				throw new AuthenticationError(
+					"You don't have permissions for this action"
+				);
+			}
 			const {
 				id,
 				category_id,
@@ -73,6 +94,7 @@ const productResolvers = {
 				shipping_cost,
 				discount,
 			} = input;
+			if (name.length < 3) throw new UserInputError("Name is too short");
 
 			const arrOfImgs =
 				input.img_id.length > 0
@@ -93,7 +115,7 @@ const productResolvers = {
 				images: arrOfImgs,
 			};
 
-			return prisma.product
+			const product = await prisma.product
 				.update({
 					where: { id: id },
 					data: updatedProduct, // probably wrong #any
@@ -101,8 +123,24 @@ const productResolvers = {
 				.catch(() => {
 					input.img_id.forEach((id) => cloudinary.uploader.destroy(id!));
 				});
+			await prisma.category.update({
+				where: { id: category_id },
+				data: {
+					companies: { connect: { id: company_id } },
+				},
+			});
+			return product;
 		},
-		deleteProduct: async (parent: any, { id }: { id: number }) => {
+		deleteProduct: async (
+			parent: any,
+			{ id }: { id: number },
+			{ req }: { req: Request }
+		) => {
+			if (!req.session.user?.role.includes(Role.ADMIN)) {
+				throw new AuthenticationError(
+					"You don't have permissions for this action"
+				);
+			}
 			const data = await prisma.product.delete({ where: { id: id } });
 
 			if (data) {
